change beneficiary tests
