+++ integration_tests/src/tests/mod.rs
@@ -4,7 +4,8 @@ mod batch_onboarding;
 pub use batch_onboarding::*;
 mod batch_onboarding_deals_test;
 pub use batch_onboarding_deals_test::*;
-// mod change_beneficiary_test;
+mod change_beneficiary_test;
+pub use change_beneficiary_test::*;
 // mod commit_post_test;
 // mod datacap_tests;
 // mod evm_test;
